% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_kgraph.R
\name{build_kgraph_from_fit}
\alias{build_kgraph_from_fit}
\title{Build a knowledge graph from a fit object}
\usage{
build_kgraph_from_fit(selected_concepts, m_embeds, fit_kg, ...)
}
\arguments{
\item{selected_concepts}{Concepts of interest}

\item{m_embeds}{Embeddings matrix}

\item{fit_kg}{Fit object}

\item{...}{Passed to build_kgraph}
}
\value{
Knowledge graph, list of slots df_nodes and df_links
}
\description{
Computes similarities for nodes of interest on the fly to avoid having to
deal with very large similarity matrices when number of features is large.
}
