% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_phecode_pairs.R
\docType{data}
\name{df_phecode_pairs}
\alias{df_phecode_pairs}
\title{A dataset containing Phecode pairs}
\format{
A dataframe with 3288 rows and 4 columns.
}
\usage{
data("df_phecode_pairs")
}
\description{
The dataframe provides clinician-curated pairs of related of
             medical concepts, useful to evaluate the performance of a
             machine learning model. It's an extract of the PrimeKG database
             (see vignette for URL).
}
\details{
Each row defines a relationship between two Phecodes, along with their textual
descriptions.
}
