% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_pval.R
\docType{data}
\name{df_pval}
\alias{df_pval}
\title{A dataset containing GWAS p-values}
\format{
A dataframe with 364 rows and 3 columns
}
\usage{
data("df_pval")
}
\description{
This dataframe provides association scores between SNPs and
             mantal health-related phenotypes.
}
\details{
Each row defines an association between a SNP and a phenotype. Downloaded
from GWAS Catalog at https://www.ebi.ac.uk/gwas/efotraits/EFO_0007623.
}
\examples{
\dontrun{
data('df_pval')

kg_obj = build_kgraph('EFO_0007623', df_pval)              
}

}
