% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_pval_dict.R
\docType{data}
\name{df_pval_dict}
\alias{df_pval_dict}
\title{A dictionary for the df_pval object}
\format{
A dataframe with 333 rows and 4 columns.
}
\usage{
data("df_pval_dict")
}
\description{
Dataframe with columns id (for the phenotype or SNP identifier),
             desc (textual description), group, and color
}
\details{
Row IDs correspond to the identifiers found in columns concept1 and concept2
of the df_pval object.
}
\examples{
data('df_pval')
data('df_pval_dict')

kg_obj = build_kgraph(c('EFO_0007623', 'EFO_0007624'), df_pval, df_pval_dict)              

}
