% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_donut.R
\name{donut}
\alias{donut}
\title{Donut Chart}
\usage{
donut(data, color = NULL)
}
\arguments{
\item{data}{a data.frame with these columns: category, count (if more they will be ignored)}

\item{color}{a vector defining the colors to be applied. If color has length = 1 (single value), the
ColorBrewer palette = color will be used. If length > 1, the vector is used as a color palette.}
}
\value{
a plot (from ggplot2)
}
\description{
Donut Chart
}
\examples{

# Create test data.
mydata <- data.frame(
  category=c("Draft", "Planned", "InWork", "Done"),
  count=c(10, 20, 30, 40))

# Create color palette
mycolor <- c("#999999", "#E69F00", "#56B4E9", "#999999")

# donut with given colors
donut(mydata, mycolor)

# donut with ColorBrewer palette
donut(mydata, 5)

# donut with ColorBrewer palette
donut(mydata, "PRGn")

# donut with default ColorBrewer palette (PiYG)
donut(mydata)

}
