% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_embeds.R
\name{gen_df_notpairs}
\alias{gen_df_notpairs}
\title{Generate null pairs}
\usage{
gen_df_notpairs(
  ids,
  df_pairs = NULL,
  n_notpairs = if (is.null(df_pairs)) 1000 else nrow(df_pairs)
)
}
\arguments{
\item{ids}{Identifiers to sample from}

\item{df_pairs}{Known pairs data frame, to make sure no null pairs are in}

\item{n_notpairs}{Direct parameter to set number of null pairs returned,
bypasses parameter type.}
}
\value{
Data frame with columns concept1, concept2, weight
}
\description{
Generate null pairs
}
