% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_funs.R
\name{get_sgraph}
\alias{get_sgraph}
\title{Wrapper to build a sgraph object fromk a kgraph object}
\usage{
get_sgraph(
  l_graph,
  colors_mapping = NULL,
  label_attrs = "label",
  igraph = NULL,
  ...
)
}
\arguments{
\item{l_graph}{List of df_nodes and df_links dataframes}

\item{colors_mapping}{Output of get_colors_map}

\item{label_attrs}{Column name of df_nodes that will be displayed}

\item{igraph}{Intermediary igraph object, if already computed}

\item{...}{Passed to sgraph::sgraph_clusters}
}
\value{
Sgraph htmlwidget object
}
\description{
Wrapper to build a sgraph object fromk a kgraph object
}
