% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_circular_barplot.R
\name{k_circular_barplot}
\alias{k_circular_barplot}
\title{Title}
\usage{
k_circular_barplot(data)
}
\arguments{
\item{data}{a data.frame of the data to plot, with columns (xxx). Other columns will be ignored.}
}
\value{
a plot (from @ggplot function)
}
\description{
Title
}
