% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_exploded_radar.R
\name{k_exploded_radar}
\alias{k_exploded_radar}
\title{Title}
\usage{
k_exploded_radar(
  data,
  title = NULL,
  width = 0.25,
  color = "#C06C84",
  alpha = c(0.2, 0.8),
  progress_line = NULL,
  progress_line_type = "dashed",
  progress_line_color = "black",
  progress_point = NULL,
  progress_point_size = 3,
  progress_point_color = "black"
)
}
\arguments{
\item{data}{a data.frame of the data to plot, with columns (category, nb, progress). Other columns will be ignored.}

\item{title}{a label to set as title of the plot. NULL will display no title (DEFAULT).}

\item{width}{the desired width for the radar area. It will be used as an offset from the center line, so each bar will be 2*width thick. DEFAULT = 0.25}

\item{color}{the reference color to fill the sector. DEFAULT = "#C06C84"}

\item{alpha}{a vector of length 2 c(min, max). Will be used to build a sequence of @color passed to scale_fill_manual. DEFAULT = c(0.2, 0.8).}

\item{progress_line}{a vector to define which sector should have a progress line in it. Will be used to subset @data columns. DEFAULT = NULL.}

\item{progress_line_type}{the desired type of line. DEFAULT = "dashed".}

\item{progress_line_color}{the desired color of line. DEFAULT = "black".}

\item{progress_point}{a vector to define which sector should have a progress point in it. Will be used to subset @data columns. DEFAULT = NULL.}

\item{progress_point_size}{the desired size for the point. DEFAULT = 3.}

\item{progress_point_color}{the desired color for the point. DEFAULT = "black.}
}
\value{
a plot (from @ggplot function)
}
\description{
Title
}
