% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_embeds.R
\docType{data}
\name{m_embeds}
\alias{m_embeds}
\title{A dataset containing medical word embeddings}
\format{
A matrix with 1122 rows and 100 columns.
}
\usage{
data("m_embeds")
}
\description{
The embedding matrix has been fitted using Glove word
             embeddings on 1,700 open-access publications related to
             mental health.
}
\details{
Each row is the embedding vector of a CUI in 100 Glove dimensions.
}
\examples{
\dontrun{
data('m_embeds')

fit_kg = fit_embeds_kg(m_embeds, 'cosine')
}

}
