% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_embeds.R
\name{project_pairs}
\alias{project_pairs}
\title{Predict known pairs}
\usage{
project_pairs(m_simi, threshold)
}
\arguments{
\item{m_simi}{Similarity matrix}

\item{threshold}{Similarity value threshold}
}
\value{
Data frame with columns concept1, concept2, weight
}
\description{
Predict known pairs
}
