% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts_df_example.R
\name{reshape_multiple_traits}
\alias{reshape_multiple_traits}
\title{Reshape multiple traits in example data}
\usage{
reshape_multiple_traits(df_pval)
}
\arguments{
\item{df_pval}{Data frame of p-values}
}
\value{
Reshaped data frame
}
\description{
Reshape multiple traits in example data
}
