% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts_df_example.R
\name{reshape_multiple_traits_dict}
\alias{reshape_multiple_traits_dict}
\title{Reshape multiple traits in example data dictionary}
\usage{
reshape_multiple_traits_dict(df_dict)
}
\arguments{
\item{df_dict}{Data frame of p-values dictionary}
}
\value{
Reshaped data frame
}
\description{
Reshape multiple traits in example data dictionary
}
